requirements.txt
pybedtools==0.7.10
pyvcf
datetime
